Add set cookie test
