+++ apps/question-bank/src/main/java/gov/cdc/nbs/questionbank/page/classic/preview/ClassicPagePreviewRedirector.java
@@ -1,15 +1,12 @@
 package gov.cdc.nbs.questionbank.page.classic.preview;
 
 import java.net.URI;
-import javax.servlet.http.HttpServletRequest;
 import org.springframework.http.HttpHeaders;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.security.access.prepost.PreAuthorize;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
-import org.springframework.web.bind.annotation.RequestMapping;
-import org.springframework.web.bind.annotation.RequestMethod;
 import org.springframework.web.bind.annotation.RestController;
 import org.springframework.web.util.UriComponentsBuilder;
 import gov.cdc.nbs.questionbank.page.classic.ClassicPreviewPagePreparer;
@@ -22,8 +19,6 @@ import springfox.documentation.annotations.ApiIgnore;
 public class ClassicPagePreviewRedirector {
 
   private static final String PREVIEW = "/nbs/PreviewPage.do";
-  private static final String EDIT_PAGE = "/page-builder/edit/page/";
-  private static final String PAGE_LIBRARY = "/page-builder/manage/pages";
 
   private final ClassicPreviewPagePreparer preparer;
 
@@ -46,18 +41,6 @@ public class ClassicPagePreviewRedirector {
     return ResponseEntity.status(HttpStatus.FOUND)
         .header(HttpHeaders.LOCATION, location.toString())
         .headers(pageCookie.apply())
-        .build();
-  }
-
-  @RequestMapping(path = "/api/v1/pages/return", method = {RequestMethod.GET, RequestMethod.POST})
-  ResponseEntity<Void> returnToEdit(final HttpServletRequest request) {
-    String location = ReturningPageCookie.resolve(request.getCookies())
-        .map(c -> EDIT_PAGE + c.page())
-        .orElse(PAGE_LIBRARY);
-
-    return ResponseEntity.status(HttpStatus.FOUND)
-        .header(HttpHeaders.LOCATION, location)
-        .headers(ReturningPageCookie.empty().apply())
         .build();
   }
